django-backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
django-redis==4.12.1
djangorestframework==3.12.4
idna==2.10
pytz==2021.1
redis==3.5.3
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
